backends/requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
astunparse==1.6.3
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmdstanpy==1.0.8
colorama==0.4.6
contourpy==1.0.6
convertdate==2.4.0
cryptography==38.0.3
cycler==0.11.0
ephem==4.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
github3.py==3.2.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.17
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
libclang==14.0.6
LunarCalendar==0.0.9
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
setuptools-git==1.2
six==1.16.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0
